add: adding prepaid cards loader animation and buttons
